+++ src/helpers/menu.builder.helper.ts
@@ -42,10 +42,8 @@ export const stickerMenu = (botInfo : Bot)=>{
 *|*► *${prefix}s* - Imagem/vídeo para sticker
 *|*► *${prefix}s* 1 - Imagem para sticker (circular)
 *|*► *${prefix}s* 2 - Imagem para sticker (sem corte)
-*|*► *${prefix}snome* pack, autor - Renomeia sticker
 *|*► *${prefix}simg* - Sticker para imagem
 *|*► *${prefix}ssf* - Imagem para sticker (sem fundo)
-*|*► *${prefix}emojimix* 💩+😀 - Emoji para sticker
 *|*
 _*M ᴏ ᴅ ᴅ ᴇ ᴅ B ʏ J ᴏ ɴ ɪ ʏ & P ᴀ ᴜ ʟ ᴏ*_ `
 }
@@ -73,6 +71,7 @@ export const downloadMenu = (botInfo : Bot)=>{
 *|*━━━━ Guia ❔: *${prefix}comando* guia
 *|*
 *|*━━━━ ✦ 📥 *DOWNLOADS* ✦
+*|*► *${prefix}d* link - Download automático de mídias
 *|*► *${prefix}play* nome - Áudio do Youtube
 *|*► *${prefix}yt* nome - Vídeo do Youtube
 *|*► *${prefix}fb* link - Vídeo do Facebook
@@ -84,7 +83,7 @@ export const downloadMenu = (botInfo : Bot)=>{
 _*M ᴏ ᴅ ᴅ ᴇ ᴅ B ʏ J ᴏ ɴ ɪ ʏ & P ᴀ ᴜ ʟ ᴏ*_ `
 }
 
-// MENU - UTILIDADE
+// MENU - UTILIDADE (USUÁRIOS COMUNS)
 export const utilityMenu = (botInfo : Bot)=>{
     let {name, prefix} = botInfo
     return `*|*━━━ ✦ *🤖 ${name?.trim()}* ✦
@@ -92,38 +91,25 @@ export const utilityMenu = (botInfo : Bot)=>{
 *|*━━━━ Guia ❔: *${prefix}comando* guia
 *|*
 *|*━━━━ ✦ ⚒️ *UTILITÁRIOS* ✦
-*|*
-*|*━━ ✦ 🔎 *CONSULTAS/TEXTO* ✦
-*|*► *${prefix}steamverde* jogo - Pesquisa de jogos "alternativos"
-*|*► *${prefix}brasileirao* - Tabela do Brasileirão
-*|*► *${prefix}animes* - Últimos lançamentos de animes
-*|*► *${prefix}mangas* - Últimos lançamentos de mangás
-*|*► *${prefix}filmes* - Tendências atuais de filmes
-*|*► *${prefix}series* - Tendências atuais de séries
-*|*► *${prefix}encurtar* link - Encurtador de link
-*|*► *${prefix}letra* musica - Letra de música
-*|*► *${prefix}traduz* idioma texto - Tradutor de texto
-*|*► *${prefix}pesquisa* texto - Pesquisa do Google
-*|*► *${prefix}clima* cidade - Previsão do tempo
-*|*► *${prefix}noticias* - Notícias atuais
-*|*► *${prefix}moeda* tipo valor - Conversor de moeda
-*|*► *${prefix}calc* expressao - Calculadora
-*|*► *${prefix}ddd* - Informação do DDD
-*|*► *${prefix}tabela* - Tabela de caracteres
-*|*
-*|*━━ ✦ 🔊 *AUDIO* ✦
-*|*► *${prefix}ouvir* - Áudio para texto
+*|*► *${prefix}save* - Salvar status
 *|*► *${prefix}audio* - Extrai áudio de um video
-*|*► *${prefix}efeitoaudio* tipo - Adiciona efeito no áudio
-*|*► *${prefix}voz* pt texto - Texto para áudio
+*|*► *${prefix}audios* - Lista de áudios disponíveis
 *|*
-*|*━━ ✦ 🖼️ *IMAGENS* ✦
-*|*► *${prefix}upimg* - Upload de imagem
-*|*► *${prefix}rbg* - Removedor de fundo
+_*M ᴏ ᴅ ᴅ ᴇ ᴅ B ʏ J ᴏ ɴ ɪ ʏ & P ᴀ ᴜ ʟ ᴏ*_ `
+}
+
+// MENU - UTILIDADE (DONO DO BOT)
+export const utilityMenuOwner = (botInfo : Bot)=>{
+    let {name, prefix} = botInfo
+    return `*|*━━━ ✦ *🤖 ${name?.trim()}* ✦
 *|*
-*|*━━ ✦ ❔ *RECONHECIMENTO* ✦
-*|*► *${prefix}qualmusica* - Reconhecimento de música
-*|*► *${prefix}qualanime* - Reconhecimento de anime
+*|*━━━━ Guia ❔: *${prefix}comando* guia
+*|*
+*|*━━━━ ✦ ⚒️ *UTILITÁRIOS* ✦
+*|*► *${prefix}revelar* - Revelar mensagem visualizar única
+*|*► *${prefix}save* - Salvar status
+*|*► *${prefix}audio* - Extrai áudio de um video
+*|*► *${prefix}audios* - Lista de áudios disponíveis
 *|*
 _*M ᴏ ᴅ ᴅ ᴇ ᴅ B ʏ J ᴏ ɴ ɪ ʏ & P ᴀ ᴜ ʟ ᴏ*_ `
 }
@@ -259,26 +245,55 @@ export const miscGroupMenu = (botInfo : Bot) =>{
 *|*
 *|*━━━━ Guia ❔: *${prefix}comando* guia
 *|*
-*|*━━ ✦ 🕹️ *JOGOS* ✦
-*|*► *${prefix}ppt* opcão - Joga pedra, papel e tesoura
-*|*► *${prefix}caracoroa* - Joga cara ou coroa
-*|*► *${prefix}roletarussa* - Joga roleta russa
+*|*━━━━ ✦ 🧩 *VARIADO* ✦
+*|*► *${prefix}vtnc* - Vai tomar no c*
 *|*
-*|*━━ ✦ 🎲 *SORTEIO* ✦
-*|*► *${prefix}sorteio* numero - Sorteia um número até esse valor.
-*|*► *${prefix}sorteiomembro* - Sorteia um membro do grupo.
+*|*━━━━ ✦ ⚙️ *ADMINISTRAÇÃO* ✦
 *|*
-*|*━━ ✦ 🧩 *ENTRETENIMENTO* ✦
-*|*► *${prefix}mascote* - Onipotente e onipresente WhatsApp Jr
-*|*► *${prefix}frase* - Frase dúvidosa do WhatsApp Jr
-*|*► *${prefix}viadometro* - Nível de viadagem
-*|*► *${prefix}detector* - Detector de mentira
-*|*► *${prefix}casal* - Escolhe um casal
-*|*► *${prefix}gadometro* - Nível de gado
-*|*► *${prefix}chance* texto - Chance de algo acontecer
-*|*► *${prefix}bafometro* - Nível de álcool
-*|*► *${prefix}top5* tema - Ranking de top 5
-*|*► *${prefix}par* @pessoa1 @pessoa2 - Nível de compatibilidade
+*|*━━ ✦ 🛠️ *GERAL* ✦
+*|*► *${prefix}info* - Informação do bot
+*|*► *${prefix}ping* - Informação do sistema
+*|*► *${prefix}bloquear* @usuario  - Bloqueia o usuário
+*|*► *${prefix}desbloquear* @usuario  - Desbloqueia o usuário
+*|*► *${prefix}listablock*  - Lista de usuários bloqueados
+*|*► *${prefix}bcgrupos* texto - Mensagem para todos os grupos
+*|*► *${prefix}desligar* - Desliga o bot
+*|*
+*|*━━ ✦ 🎨 *CUSTOMIZAÇÃO* ✦
+*|*► *${prefix}nomebot* nome - Altera nome do bot
+*|*► *${prefix}prefixo* simbolo - Altera o prefixo dos comandos
+*|*► *${prefix}fotobot* - Altera foto do bot
+*|*► *${prefix}recado* texto - Altera o texto do recado/status
+*|*
+*|*━━ ✦ 👨‍👩‍👧‍👦 *GRUPOS* ✦
+*|*► *${prefix}grupos* - Dados dos grupos atuais
+*|*► *${prefix}entrargrupo* link - Entra no grupo
+*|*
+*|*━━ ✦ 👤 *USUÁRIOS* ✦
+*|*► *${prefix}usuario* @usuario - Dados do usuário
+*|*
+*|*━━ ✦ ⭐ *ADMINS* ✦
+*|*► *${prefix}admins* - Administradores do bot
+*|*► *${prefix}addadmin* - Promove a admin do bot
+*|*► *${prefix}rmadmin* - Rebaixa a usuário do bot
+*|*
+*|*━━━━ ✦ 🧰  *RECURSOS* ✦
+*|*
+*|*━ ✦ 🏞️  *AUTO-STICKER PRIVADO* ✦
+*|*► *${prefix}autostickerpv* - Ativa/desativa a criação automática de stickers no privado
+*|*
+*|*━ ✦ 🔒 *BLOQUEIO DE COMANDOS* ✦
+*|*► *${prefix}bcmdglobal* !cmd1 !cmd2 - Bloqueia os comandos globalmente
+*|*► *${prefix}dcmdglobal* !cmd1 !cmd2 - Desbloqueia os comandos globalmente
+*|*
+*|*━ ✦ ⭐ *MODO ADMIN* ✦
+*|*► *${prefix}modoadmin* - Ativa/desativa o modo para apenas admins do bot usarem comandos
+*|*
+*|*━ ✦ ⏳ *TAXA DE COMANDOS* ✦
+*|*► *${prefix}taxacomandos* numero - Ativa/desativa a taxa de comandos por minuto
+*|*
+*|*━ ✦ 📩 *MENSAGENS PRIVADAS* ✦
+*|*► *${prefix}comandospv* - Ativa/desativa os comandos em mensagens privadas
 *|*
 _*M ᴏ ᴅ ᴅ ᴇ ᴅ B ʏ J ᴏ ɴ ɪ ʏ & P ᴀ ᴜ ʟ ᴏ*_ `
 }
@@ -339,4 +354,3 @@ export const adminMenu = (botInfo : Bot)=>{
 *|*
 _*M ᴏ ᴅ ᴅ ᴇ ᴅ B ʏ J ᴏ ɴ ɪ ʏ & P ᴀ ᴜ ʟ ᴏ*_ `
 }
-