+++ webpack.entry.js
@@ -1,1 +1,0 @@
-console.log('Webpack smoke build executed.');